+++ nukkit/src/main/java/io/github/nemo_64/chatinput/nukkit/impl/NkktSender.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.github.nemo_64.nukkit.impl;
+package io.github.nemo_64.chatinput.nukkit.impl;
 
 import cn.nukkit.Player;
 import io.github.nemo_64.chatinput.Sender;
